Funcion del tablero que inicializa el color y el tipo de las piezas
He hecho una prueba que muestra la inicializacion al copilar. Tambien he corregido algunas cosas como el tipo del atributo "tipo" a Tipo.

+++ src/alfil.h
@@ -2,6 +2,6 @@
 #include "pieza.h"
 
 class Alfil : public Pieza{
-	void set_tipo() { tipo = static_cast<int>(Tipo::ALFIL); }
+	void set_tipo() { tipo = Tipo::ALFIL; }
 	
 };
