+++ src/caballo.h
@@ -2,6 +2,6 @@
 #include "pieza.h"
 
 class Caballo : public Pieza {
-	void set_tipo() { tipo = static_cast<int>(Tipo::CABALLO); }
+	void set_tipo() { tipo = Tipo::CABALLO; }
 	
 };
