+++ src/peon.h
@@ -2,6 +2,6 @@
 #include "pieza.h"
 
 class Peon : public Pieza {
-	void set_tipo() { tipo = static_cast<int>(Tipo::PEON); }
+	void set_tipo() { tipo = Tipo::PEON; }
 	
 };
